state 관련 파일 분리
