+++ .eslintrc.json
@@ -25,7 +25,7 @@
     "import/order": [
       "warn",
       {
-        "groups": ["builtin", "external", ["parent", "sibling"], "index"],
+        "groups": ["builtin", "external", ["parent", "sibling", "index"]],
         "newlines-between": "always",
         "alphabetize": {
           "order": "asc",
