+++ src/states/models.ts
@@ -1,0 +1,14 @@
+import { Tier } from "../Tier";
+
+export type TitleJSON = {
+  tier: Tier;
+  titleId: number;
+  mute: boolean;
+  articles: ArticleJSON[];
+};
+
+export type ArticleJSON = {
+  no: number;
+  readAt?: string;
+  syncAt?: string;
+};
